[Finished #53766277] Debugging is now done directly from the AbstractCommunication class. Debug.h and Debug.cpp removed.

+++ GlobalReceiver.cpp
@@ -34,7 +34,7 @@ void GlobalReceiver::process() {
 		sprintf(reply, "Meta (%s)", input.c_str());
 		break;
 	default:
-		sprintf(reply, "??? (%s): %x", input.c_str(), action);
+		sprintf(reply, "??? (%s): %x", input.c_str(), (int)action);
 	}
 
 	com->send(String(reply));
